Share code for completeBoundaryWithStyles
